Move table classes to separate file
